findlib: Move object files to -dbg package.
Not quite sure this belongs in -dev anyway, but maybe the findlib will
re-use the object files if they're there and compatible? (... just for
bytecode?).

Anyway, from OE package.bbclass perspective, rpmdeps will find runtime
dependencies on .cm[ia] files which cannot be satisfied since this layer
"cross-compile" and does not provide a target ocamlrun for the target
(it could...). This makes it "debug" feature, on which
runtime-dependencies are not processed, which makes OE happy, for now.

E,g:
/usr/lib/ocaml/site-lib/stdext/stdext.cma contained in package xenclient-toolstack-dev requires ocaml(Scanf), but no providers found in RDEPENDS_xenclient-toolstack-dev? [file-rdeps]
Would be solved by:
RDEPENDS_${PN}-dev = "<ocamlrun-provider>"
... should that be a thing one day.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>