ocaml: Honor cross build environment for -cross.
With this change, the native compiler is re-used in the -cross recipe to
build a cross build environment and used as cross-compiler in the
toolchain later deployed for ocaml recipes.

OpenXT ocaml components used to link by sheer luck because libraries
(glibc 2.13-2.24) happened to be backward compatible for the use-case.
This is no longer the case with 2.28+, only revealing that the cross
build environment was invalid.

Dynamic libraries, more so OCaml hybrid libraries, do not play nice with
cross-compilation and ocamlmklib will end up linking with the host
environment (e.g, libc), eventually running into missing symbols.

Fall-back on static linking, which most cross-toolchain of OCaml seems
to end up doing.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>